+++ valeur.h
@@ -8,13 +8,13 @@ class valeur
 
 friend std::ostream & operator << (std::ostream & os, const valeur & valeur);
 public:
-    valeur () : nb(0), lien(0.0), calc(0) {}
-    valeur (int w_nb) : nb(w_nb), lien(0.0), calc(0) {}
+    valeur () : nb(0.0), lien(0.0), calc(0) {}
+    valeur (double w_nb) : nb(w_nb), lien(0.0), calc(0) {}
     ~valeur () {}
 
 public:
-    unsigned short int nb;
-    float lien;
+    double nb;
+    double lien;
     bool calc ;
 }
 ;
